+++ oinkbrew/StatusMessage.cpp
@@ -29,7 +29,6 @@
 #include "Settings.h"
 #include "spark_wiring_udp.h"
 #include "spark_wiring_wifi.h"
-#include "Version.h"
 
 /*******************************************************************************
  * Function Name  : send