+++ screens/home_screen.h
@@ -1,0 +1,43 @@
+/**
+ ******************************************************************************
+ * @file    home_screen.h
+ * @authors Thomas Trageser
+ * @version V0.1
+ * @date    2015-03-25
+ * @brief   Oink Brew Spark Core Firmware
+ ******************************************************************************
+  Copyright (c) 2015 Oink Brew;  All rights reserved.
+
+  This program is free software; you can redistribute it and/or
+  modify it under the terms of the GNU Lesser General Public
+  License as published by the Free Software Foundation, either
+  version 3 of the License, or (at your option) any later version.
+
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+  Lesser General Public License for more details.
+
+  You should have received a copy of the GNU Lesser General Public
+  License along with this program; if not, see <http://www.gnu.org/licenses/>.
+ ******************************************************************************
+ */
+
+
+#ifndef HOME_SCREEN_H
+#define	HOME_SCREEN_H
+
+#ifdef	__cplusplus
+extern "C" {
+#endif
+
+    
+D4D_EXTERN_SCREEN(screen_home);
+
+
+#ifdef	__cplusplus
+}
+#endif
+
+#endif	/* HOME_SCREEN_H */
+