+++ CHANGELOG.md
@@ -4,4 +4,6 @@ v0.1
 ----
 - startup check if WiFi connection and changes to listen mode if not
 - udp broadcast status message to make itself known to the Oink Brew web application
-- Application start send 1 buzz when starting and 2 buzz when startup is finished
+- Application start shows startup screen with progress information
+- Application start will show a calibration screen if no calibration settings found in EEPROM
+